Use quint32 instead of unsigned long int as the return type for idleTime
unsigned long int doesn't work so well with QVariant, so let's use something
that's more compatible. :)

svn path=/branches/work/kopete/dev-4.2/; revision=832624

+++ libkopete/kopetemetacontact.cpp
@@ -414,7 +414,7 @@ Contact *MetaContact::execute()
 	return 0L;
 }
 
-unsigned long int MetaContact::idleTime() const
+quint32 MetaContact::idleTime() const
 {
 	unsigned long int time = 0;
 	QListIterator<Contact *> it( d->contacts );
